% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_transplant_success.R
\name{check_transplant_success}
\alias{check_transplant_success}
\title{Build a community matrix to simulate community transplantation}
\usage{
check_transplant_success(
  donor,
  transplant,
  newtaxa.only = FALSE,
  tidy = TRUE,
  keep.all.taxa = TRUE
)
}
\arguments{
\item{donor}{Numeric matrix. The donor community, typically created with make_donor_community()}

\item{transplant}{Numeric matrix. The final community that is the result of a transplant_w_() function.}

\item{newtaxa.only}{Logical. If newtaxa.only=TRUE, only novel taxa that persisted after transplantation will be retained.}

\item{tidy}{Logical. If tidy=TRUE, the function will return a 'tidy' data set, as opposed to a 'wide' one. The tidy version has columns named c("timepoint","sample","taxon","relative_abundance"). Default=TRUE.}

\item{keep.all.taxa}{Logical. If keep.all.taxa=TRUE, all overlapping taxa between the donor and recipient communities will be used to calculate relative abundance values, not just the taxa that are eventually kept. Default=FALSE.}
}
\value{
Either a 'wide' or 'long' data frame with relative abundance values from the original donor community and the associated values for those taxa after transplantation.
}
\description{
Convenience funtion that gathers up donor microbiome community relative abundances before and after transplanting
Returns a data frame that can be easily used to examine the success of microbial community establishement
}
\examples{
comm <- build_even_community(n.taxa = 100,n.samples = 44,n.reads = 3000, taxa.sd = 30)
donor <- build_donor_community(resident.comm = comm, n.transplant.taxa = 30,overlap = .75)
transplant <- transplant_w_niche_occupation(comm,donor)
check_transplant_success(donor,transplant,newtaxa.only=TRUE)

}
