% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_competition.R
\name{simulate_competition}
\alias{simulate_competition}
\title{Simulates the actions of niche overlap and differential fitness to alter a taxa abundance table.}
\usage{
simulate_competition(
  dat,
  n.niches = ncol(dat)/2,
  niche.distr = "normal",
  niche.separation = n.niches^2
)
}
\arguments{
\item{dat}{Numeric matrix. The starting community matrix that will be edited.}

\item{n.niches}{Positive whole numeric vector of length 1. The number of niches to simulate. All taxa will be assigned to one of these niches. Default=ncol(dat)/2.}

\item{niche.distr}{Character vector of length 1. Must currently be in c("normal"). The distribution from which to draw niche overlap values. Default='normal'.}

\item{niche.separation}{Numeric vector of length 1. Indicates the standard deviation for the initial seeds that generate niches. Larger values lead to less niche overlap. Default=n.niches^2.}
}
\value{
Taxon abundance matrix with samples as rows and taxa as columns. class='matrix'. Abundance values are updated to reflect all competitive outcomes.
}
\description{
Manipulates a community matrix with samples as rows and taxa as columns. Taxa are assigned random ranges within a number of niche distributions. Taxa are also assigned random "fitness" values. Each possible pairing of taxa are matched and competition is simulated based on any niche overlap and differential fitness. Taxon abundances are altered accordingly for both taxa.
}
\examples{
comm <- build_even_community(n.taxa = 100, n.samples = 44, n.reads = 3000, taxa.sd = 30)
sim <- simulate_competition(comm, n.niches=15, niche.separation=5)
colSums(comm);colSums(sim)

}
