% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transplant_w_stochasticity.R
\name{transplant_w_stochasticity}
\alias{transplant_w_stochasticity}
\title{Simulate community transplantation with stochastisicity in donor taxa abundances}
\usage{
transplant_w_stochasticity(recipient, donor, stochasticity = 0.05)
}
\arguments{
\item{recipient}{Numeric matrix. The starting community matrix that represents the resident community in N samples.}

\item{donor}{Numeric matrix. The donor community matrix that represents the donor community. Typically, the result of build_donor_community().}

\item{stochasticity}{Positive numeric vector of length 1, between 0 and 1. The level of random noise to add to novel taxa abundances. If set to 0, no random noise will be added. Default=.05}
}
\value{
Taxon abundance matrix with samples as rows and taxa as columns. class='matrix'.
}
\description{
Manipulates a community matrix with samples as rows and taxa as columns. Donor taxa that overlap with recipient taxa will be joined additively. Novel taxa will be added with random noise.
}
\examples{
recipient <- build_even_community(n.taxa = 100,n.samples = 44,n.reads = 3000, taxa.sd = 30)
donor <- build_donor_community(resident.comm = recipient,n.transplant.taxa = 10,overlap = .3)
transplant_w_stochasticity(recipient, donor, stochasticity=0.05)

}
