% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_even_community.R
\name{build_even_community}
\alias{build_even_community}
\title{Build even community matrix}
\usage{
build_even_community(
  n.taxa = 3000,
  n.samples = 100,
  n.reads = 5e+06,
  prop.dominant = 0.01,
  prop.absent = 0.2
)
}
\arguments{
\item{n.taxa}{Positive numeric vector of length 1. The number of taxa (columns) to generate. Default = 500.}

\item{n.samples}{Positive numeric vector of length 1. The number of samples (rows) to generate. Default = 100.}

\item{n.reads}{Positive numeric vector of length 1. The number of reads (observations) for the entire study. This is analogous to overall sequencing depth. Default = 5,000,000}

\item{prop.dominant}{The proportion of taxa that are dominant; represented as > 1 percent relative abundance. Default = .01}

\item{prop.absent}{The rough proportion of taxa that are missing from a given sample. Default = 0.2}
}
\value{
Taxon abundance matrix with samples as rows and taxa as columns. class='matrix'
}
\description{
Builds a community matrix with samples as rows and taxa as columns.
Taxa abundances are distributed in decreasing value after the dominant taxa are apportioned their shares.
}
\examples{
comm <- build_even_community(n.taxa = 3000, n.samples = 100, n.reads = 500000, prop.dominant = .02)

}
